refactor(admin): split dashboard route panels
Move the dashboard route shell primitives and async data panels into local modules so page.tsx owns only rendered timestamp, layout composition, and suspense boundaries.

Verified with pnpm exec prettier --check apps/admin/src/app/page.tsx apps/admin/src/app/dashboard-panels.tsx apps/admin/src/app/dashboard-sections.tsx; pnpm --filter @anipotts/admin typecheck; pnpm --filter @anipotts/admin lint; pnpm --filter @anipotts/admin build; git diff --check -- apps/admin/src/app/page.tsx apps/admin/src/app/dashboard-panels.tsx apps/admin/src/app/dashboard-sections.tsx; pnpm validate; pnpm dlx knip --no-exit-code.
